Added SimpleAPIView and refactor APIView
* Added View as base
* Added tests
