+++ esmerald/routing/generics/__init__.py
@@ -1,0 +1,6 @@
+from .views import APIView, SimpleAPIView
+
+__all__ = [
+    "APIView",
+    "SimpleAPIView",
+]
