+++ esmerald/typing.py
@@ -1,7 +1,5 @@
-from enum import Enum
-from typing import Any, Callable, Dict, Type, TypeVar, Union
+from typing import Any, Callable, Type, TypeVar
 
-from pydantic import BaseModel
 from pydantic_core import PydanticUndefined
 
 T = TypeVar("T")
@@ -14,4 +12,3 @@ class Void:
 VoidType = Type[Void]
 AnyCallable = Callable[..., Any]
 Undefined = PydanticUndefined
-ModelMap = Dict[Union[Type[BaseModel], Type[Enum]], str]
